django_backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
mysqlclient>=2.2
boto3>=1.34
django-storages>=1.14
resend>=0.7
python-dotenv>=1.0
gunicorn>=21.2
Pillow>=10.0
django-filter>=23.5
beautifulsoup4
lxml
